Eliminated an `implicit conversion of float to int' compilation warning from crossbld.cpp.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@7592 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/crossbld/crossbld.cpp
@@ -335,7 +335,7 @@ bool csCrossBuilder::BuildSpriteFactory (iModelData *Data,
 
 	  int FrameIndex = Frames.Find (ver);
 	  CS_ASSERT (FrameIndex != -1);
-	  spract->AddFrame (tgt->GetFrame (FrameIndex), Delay * 1000);
+	  spract->AddFrame (tgt->GetFrame (FrameIndex), int(Delay * 1000));
 	}
       }
     }
